agent/pyproject.toml
[project]
name = "digital-human-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.1",
  "livekit-agents[tavus]~=1.3",
  "livekit-plugins-openai~=1.0",
]

[tool.uv]
dev-dependencies = []
